requirements.txt
aiogram
python-dotenv
gspread
gspread_dataframe
pandas
sqlalchemy
aiosqlite
greenlet
yookassa